requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
beautifulsoup4==4.12.3
openai==1.50.0
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.0
